requirements.txt
-i https://pypi.org/simple
boto3==1.9.253
botocore==1.12.253
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
docutils==0.15.2
idna==2.8
jmespath==0.9.4
lockfile==0.12.2
luigi==2.8.9
psycopg2==2.8.3
python-daemon==2.1.2
python-dateutil==2.8.0 ; python_version >= '2.7'
python-decouple==3.1
python-json-logger==0.1.11
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlalchemy==1.3.10
tornado==5.1.1
urllib3==1.25.6 ; python_version >= '3.4'
